+++ src/main/java/com/excilys/computerdatabase/model/Company.java
@@ -1,14 +1,14 @@
 package com.excilys.computerdatabase.model;
 
 public class Company {
-	private Integer id;
+	private Long id;
 	private String name;
 		
 	public Company(String name) {
 		this.name = name;
 	}
 	
-	public Company(Integer id, String name){
+	public Company(Long id, String name){
 		this.id = id;
 		this.name = name;
 	}
@@ -17,7 +17,7 @@ public class Company {
 		return name;
 	}
 	
-	public Integer getId(){
+	public Long getId(){
 		return id;
 	}
 }
